completed ECS, made simple example for ECS, made new Bitset tests

+++ include/mercer/ECS/ComponentBitmask.hpp
@@ -28,7 +28,7 @@ SOFTWARE.
 
 namespace mercer {
 
-using BitmaskIndexType = unsigned long long;
+using BitmaskIndexType = int;
 using Bitmask = Bitset<BitmaskIndexType>;
 
 }
